Finished authentication and middlewares
