+++ src/server.js
@@ -5,7 +5,9 @@ const middlewares = require('./middlewares');
 const mongoose = require('mongoose');
 require('dotenv').config();
 
-mongoose.connect('mongodb+srv://doge:doge@cluster0-iivj7.mongodb.net/test?retryWrites=true&w=majority',{
+mongoose.connect(
+  'mongodb+srv://' + process.env.MONGO_NAME + ':' + process.env.MONGO_PASS +
+  '@cluster0-iivj7.mongodb.net/test?retryWrites=true&w=majority', {
   useNewUrlParser: true,
 });
 
@@ -17,7 +19,7 @@ const server = new GraphQLServer({
   middlewares: [middlewares.Permissions],
   context: req => ({
     ...req,
-    token: middlewares.Token(req)
+    token: middlewares.token(req)
   })
 });
